Move Core tests into correct namespace.
